Avoid integer underflow in pagination by making "remaining" a huge signed type.
The method already seems to treat it like it can go under zero with the while condition

+++ Modio/Clients/SearchClient.cs
@@ -79,7 +79,8 @@ public class SearchClient<T> : ApiClient where T : class
     public async IAsyncEnumerable<IReadOnlyList<T>> ToPagedEnumerable()
     {
         var (method, path) = this.route;
-        uint? remaining = null;
+        // Huge integer type that does signed arithmetic and encompasses uint's max value
+        long? remaining = null;
         do
         {
             var req = new Request(method, path);
